+++ src/array.rs
@@ -1,5 +1,8 @@
-use super::{Error, IterError, Metadata, Miniconf, Peekable};
-use core::ops::{Deref, DerefMut};
+use super::{Error, IterError, Metadata, Miniconf};
+use core::{
+    fmt::Write,
+    ops::{Deref, DerefMut},
+};
 
 /// An array that exposes each element through their [`Miniconf`] implementation.
 ///
@@ -109,7 +112,7 @@ const fn digits(x: usize) -> usize {
 impl<T: Miniconf, const N: usize> Miniconf for Array<T, N> {
     fn set_path<'a, 'b: 'a, P, D>(&mut self, path_parts: &mut P, de: D) -> Result<(), Error>
     where
-        P: Peekable<Item = &'a str>,
+        P: Iterator<Item = &'a str>,
         D: serde::Deserializer<'b>,
     {
         let i = self.0.index(path_parts.next())?;
@@ -122,7 +125,7 @@ impl<T: Miniconf, const N: usize> Miniconf for Array<T, N> {
 
     fn get_path<'a, P, S>(&self, path_parts: &mut P, ser: S) -> Result<S::Ok, Error>
     where
-        P: Peekable<Item = &'a str>,
+        P: Iterator<Item = &'a str>,
         S: serde::Serializer,
     {
         let i = self.0.index(path_parts.next())?;
@@ -136,43 +139,31 @@ impl<T: Miniconf, const N: usize> Miniconf for Array<T, N> {
     fn metadata() -> Metadata {
         let mut meta = T::metadata();
 
-        // Unconditionally account for separator since we add it
-        // even if elements that are deferred to (`Options`)
-        // may have no further hierarchy to add and remove the separator again.
-        meta.max_length += digits(N) + 1;
+        // We add separator and index
+        meta.max_length += 1 + digits(N);
         meta.max_depth += 1;
         meta.count *= N;
 
         meta
     }
 
-    fn next_path<const TS: usize>(
-        state: &mut [usize],
-        topic: &mut heapless::String<TS>,
+    fn next_path(
+        state: &[usize],
+        depth: usize,
+        mut topic: impl Write,
         separator: char,
-    ) -> Result<bool, IterError> {
-        let original_length = topic.len();
-
-        while *state.first().ok_or(IterError::PathDepth)? < N {
-            // Add the array index and separator to the topic name.
-            topic
-                .push_str(itoa::Buffer::new().format(state[0]))
-                .and_then(|_| topic.push(separator))
-                .map_err(|_| IterError::PathLength)?;
-
-            if T::next_path(&mut state[1..], topic, separator)? {
-                return Ok(true);
+    ) -> Result<usize, IterError> {
+        match state.get(depth) {
+            Some(&i) if i < N => {
+                topic
+                    .write_char(separator)
+                    .and_then(|_| topic.write_str(itoa::Buffer::new().format(i)))
+                    .map_err(|_| IterError::Length)?;
+                T::next_path(state, depth + 1, topic, separator)
             }
-
-            // Strip off the previously prepended index, since we completed that element and need
-            // to instead check the next one.
-            topic.truncate(original_length);
-
-            state[0] += 1;
-            state[1..].fill(0);
+            Some(_) => Err(IterError::Next(depth)),
+            None => Err(IterError::Depth),
         }
-
-        Ok(false)
     }
 }
 
@@ -192,12 +183,12 @@ impl<T, const N: usize> IndexLookup for [T; N] {
 impl<T: crate::Serialize + crate::DeserializeOwned, const N: usize> Miniconf for [T; N] {
     fn set_path<'a, 'b: 'a, P, D>(&mut self, path_parts: &mut P, de: D) -> Result<(), Error>
     where
-        P: Peekable<Item = &'a str>,
+        P: Iterator<Item = &'a str>,
         D: serde::Deserializer<'b>,
     {
         let i = self.index(path_parts.next())?;
 
-        if path_parts.peek().is_some() {
+        if path_parts.next().is_some() {
             return Err(Error::PathTooLong);
         }
 
@@ -208,12 +199,12 @@ impl<T: crate::Serialize + crate::DeserializeOwned, const N: usize> Miniconf for
 
     fn get_path<'a, P, S>(&self, path_parts: &mut P, ser: S) -> Result<S::Ok, Error>
     where
-        P: Peekable<Item = &'a str>,
+        P: Iterator<Item = &'a str>,
         S: serde::Serializer,
     {
         let i = self.index(path_parts.next())?;
 
-        if path_parts.peek().is_some() {
+        if path_parts.next().is_some() {
             return Err(Error::PathTooLong);
         }
 
@@ -223,26 +214,28 @@ impl<T: crate::Serialize + crate::DeserializeOwned, const N: usize> Miniconf for
 
     fn metadata() -> Metadata {
         Metadata {
-            max_length: digits(N),
+            // We add separator and index
+            max_length: 1 + digits(N),
             max_depth: 1,
             count: N,
         }
     }
 
-    fn next_path<const TS: usize>(
-        state: &mut [usize],
-        path: &mut heapless::String<TS>,
-        _separator: char,
-    ) -> Result<bool, IterError> {
-        if *state.first().ok_or(IterError::PathDepth)? < N {
-            // Add the array index to the topic name.
-            path.push_str(itoa::Buffer::new().format(state[0]))
-                .map_err(|_| IterError::PathLength)?;
-
-            state[0] += 1;
-            Ok(true)
-        } else {
-            Ok(false)
+    fn next_path(
+        state: &[usize],
+        depth: usize,
+        mut path: impl Write,
+        separator: char,
+    ) -> Result<usize, IterError> {
+        match state.get(depth) {
+            Some(&i) if i < N => {
+                path.write_char(separator)
+                    .and_then(|_| path.write_str(itoa::Buffer::new().format(i)))
+                    .map_err(|_| IterError::Length)?;
+                Ok(depth)
+            }
+            Some(_) => Err(IterError::Next(depth)),
+            None => Err(IterError::Depth),
         }
     }
 }