+++ tests/option.rs
@@ -12,6 +12,13 @@ struct Settings {
 }
 
 #[test]
+fn just_option() {
+    let mut it = Option::<u32>::iter_paths::<1, 0>().unwrap();
+    assert_eq!(it.next(), Some("".into()));
+    assert_eq!(it.next(), None);
+}
+
+#[test]
 fn option_get_set_none() {
     let mut settings = Settings::default();
     let mut data = [0; 100];
@@ -19,15 +26,15 @@ fn option_get_set_none() {
     // Check that if the option is None, the value cannot be get or set.
     settings.value.take();
     assert_eq!(
-        settings.get("value_foo", &mut data),
+        settings.get("/value_foo", &mut data),
         Err(miniconf::Error::PathNotFound)
     );
     assert_eq!(
-        settings.get("value", &mut data),
+        settings.get("/value", &mut data),
         Err(miniconf::Error::PathAbsent)
     );
     assert_eq!(
-        settings.set("value/data", b"5"),
+        settings.set("/value/data", b"5"),
         Err(miniconf::Error::PathAbsent)
     );
 }
@@ -40,10 +47,10 @@ fn option_get_set_some() {
     // Check that if the option is Some, the value can be get or set.
     settings.value.replace(Inner { data: 5 });
 
-    let len = settings.get("value/data", &mut data).unwrap();
+    let len = settings.get("/value/data", &mut data).unwrap();
     assert_eq!(&data[..len], b"5");
 
-    settings.set("value/data", b"7").unwrap();
+    settings.set("/value/data", b"7").unwrap();
     assert_eq!((*settings.value).as_ref().unwrap().data, 7);
 }
 
@@ -54,13 +61,13 @@ fn option_iterate_some_none() {
     // When the value is None, it will still be iterated over as a topic but may not exist at runtime.
     settings.value.take();
     let mut iterator = Settings::iter_paths::<10, 128>().unwrap();
-    assert_eq!(iterator.next().unwrap(), "value/data");
+    assert_eq!(iterator.next().unwrap(), "/value/data");
     assert!(iterator.next().is_none());
 
     // When the value is Some, it should be iterated over.
     settings.value.replace(Inner { data: 5 });
     let mut iterator = Settings::iter_paths::<10, 128>().unwrap();
-    assert_eq!(iterator.next().unwrap(), "value/data");
+    assert_eq!(iterator.next().unwrap(), "/value/data");
     assert!(iterator.next().is_none());
 }
 
@@ -75,17 +82,17 @@ fn option_test_normal_option() {
     assert!(s.data.is_none());
 
     let mut iterator = S::iter_paths::<10, 128>().unwrap();
-    assert_eq!(iterator.next(), Some("data".into()));
+    assert_eq!(iterator.next(), Some("/data".into()));
     assert!(iterator.next().is_none());
 
-    s.set("data", b"7").unwrap();
+    s.set("/data", b"7").unwrap();
     assert_eq!(s.data, Some(7));
 
     let mut iterator = S::iter_paths::<10, 128>().unwrap();
-    assert_eq!(iterator.next(), Some("data".into()));
+    assert_eq!(iterator.next(), Some("/data".into()));
     assert!(iterator.next().is_none());
 
-    s.set("data", b"null").unwrap();
+    s.set("/data", b"null").unwrap();
     assert!(s.data.is_none());
 }
 
@@ -101,17 +108,17 @@ fn option_test_defer_option() {
     assert!(s.data.is_none());
 
     let mut iterator = S::iter_paths::<10, 128>().unwrap();
-    assert_eq!(iterator.next(), Some("data".into()));
+    assert_eq!(iterator.next(), Some("/data".into()));
     assert!(iterator.next().is_none());
 
-    assert!(s.set("data", b"7").is_err());
+    assert!(s.set("/data", b"7").is_err());
     s.data = Some(0);
-    s.set("data", b"7").unwrap();
+    s.set("/data", b"7").unwrap();
     assert_eq!(s.data, Some(7));
 
     let mut iterator = S::iter_paths::<10, 128>().unwrap();
-    assert_eq!(iterator.next(), Some("data".into()));
+    assert_eq!(iterator.next(), Some("/data".into()));
     assert!(iterator.next().is_none());
 
-    assert!(s.set("data", b"null").is_err());
+    assert!(s.set("/data", b"null").is_err());
 }